USB: move dynamic ids out of usb driver structures
The usb driver structures contain a dynamic id structure that is written
to, preventing them from being able to be constant structures.  To help
resolve this, move the dynamic id structure out of the driver and into a
separate local list, indexed off of the driver * so that all USB
subsystems can use it (i.e. usb-serial).

Overall it moves some duplicated code out of the usb-serial core as it's
already in the usb core, and now the usb dynamic id structures can be
private entirely to the usb core itself.

Cc: Johan Hovold <johan@kernel.org>
Cc: Rob Herring <robh@kernel.org>
Cc: Herve Codina <herve.codina@bootlin.com>
Cc: Alan Stern <stern@rowland.harvard.edu>
Cc: Grant Grundler <grundler@chromium.org>
Cc: Oliver Neukum <oneukum@suse.com>
Cc: Douglas Anderson <dianders@chromium.org>
Cc: Yajun Deng <yajun.deng@linux.dev>
Cc: Ivan Orlov <ivan.orlov0322@gmail.com>
Cc: "Ricardo B. Marliere" <ricardo@marliere.net>
Cc: linux-usb@vger.kernel.org
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>