container_of: switch to be a const call
Hack to switch container_of() to container_of_const() to see what
breaks.

Do NOT apply this yet, it causes massive build breakage.  The amount of
compiler splats here is amazing in the overall noise level.

But it is here so that we can work toward slowly fixing this mess up...

NACKED-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>